restored reversion for SEDTRN specl execution, fixed precision mismatch, added domain specific specl filter in HYDR

+++ HSP2/om.py
@@ -199,7 +199,7 @@ def state_om_model_run_prep(state, io_manager, siminfo):
     state['model_object_cache'] = model_object_cache
     state['model_exec_list'] = np.asarray(model_exec_list, dtype="i8") 
     if ModelObject.ops_data_type == 'ndarray':
-        state_keyvals = np.asarray(zeros(max(ModelObject.state_ix.keys()) + 1), dtype="float32")
+        state_keyvals = np.asarray(zeros(max(ModelObject.state_ix.keys()) + 1), dtype="float64")
         for ix, val in ModelObject.state_ix.items():
             state_keyvals[ix] = val
         state['state_ix'] = state_keyvals
